MI-200 Changes to the user profile user interface.
Includes resizing of the gender icons in order to fit the interface based on the resolution of the emulator.

Also includes changes in the placements of the elements on the UserProfile axml view. This is in order to fit longer names on the page, and view the gender right next to the user name, instead of below.