MI-209 Addition of a data layer, moved server logic from mapper to gateways
In order to have key value pair implementation, there must be a data
layer, currently absent. This commit contains the new data layer, and
the server logic originally found in the mappers have been moved to the
gateway package (Gateways namespace). The mappers now uses the gateway
to retrieve information from the web.

For now, the data layer may seem useless, but once local persistent
storage is implemented, the mapper first looks in the local persistent
storage for information, and only if it is not found locally, then it
will use the gateways to find the information. The Key value pairs will
be part of the local storage.