Creation of MainMenu activity
Created a new MainMenu activity that will be later used to implement the main menu page.
In addition, created the XML layout of the main menu change. Includes icons for starting a ride, preferences, user's rides, chat, user profile, and log out.
Will later add a "home" button on the bottom toolbar to redirect to the main menu.