Integrated main menu into application
Integrated the main menu page into the application and redirected the buttons on the page to the respective activities.
Also added a home button instead of the map button on the bottom toolbar that directs back to the main menu page.
Finally, added a textview greeting to the current user on the main menu page.