#80 Initial implementation of FindMatchAsyncTask
Finding a match was the most expensive operation in term of time
consumption when loading the match activity. It was also performed on
the UI Thread, which blocked the whole main thread of the application,
causing it to crash. This implementation allows the call to obtain the
matches from the server to be done asynchronously while keeping the UI
thread running. Once done, the async task calls a call back function in
the MatchActivity, which now implements the interface
IOnFindMatchCompleteCallback. The loading of the match is still slow,
but the application does not crash. A loading screen may be required
while waiting for the matches to be found.