#80 Updated the interface of the FindMatchAsyncTask to take a request instead of data to build the request
The MatchActivity needs to have access to the ID of the user's request
in order to send the confirmation to the server if the user accepts the
match. Thus, the FindMatchAsyncTask cannot build it, or if it did, it
would have to return the list of request as well as the request of the
user itself. Thus, the creation of the user request is made in the
MatchActivity rather than in the AsyncTask.