#80 GetReviewsAverageAsyncTask and Facebook profile picture conversion async
- UserGateway is now able to obtain the facebook profile picture and
convert the picture asynchronously
- UserMapper is updated to provide the support of asynchronous method
for obtaining Facebook profile picture
- The folder "Callbacks" has been added under Tasks
- New interface IOnGetReviewsAverageCompleteCallback is added to enforce
the method implementation of the callback method when an average of
reviews is finished computing
- FindMatchAsyncTask has been updated to have visibility over the new
Callbacks package
- GetReviewsAverageAsyncTask is added to obtain all feedbacks of a user
and compute the average asynchronously. Useful for statistics or to
display one representative number for the rating of a user like in the
match activity.