MI-188 Changed icons used for menu
Changed the icons used in the main menu of the application to a different style. The past buttons used were outlined with a thick black border, which isn't consistent with the rest of the graphical interfaces of our application. Instead I updated them to have a drop shadow instead of the black borders in order to be consistent with the rest of the GUI.