MI-214 Initial implementation of cache for user profile picture
Since it is relatively fast already to retrieve, convert and display a
user profile picture, we do not need to save a lot of a data for this
task. Disk storage would not benefit by a lot in comparison with calling
the facebook api to obtain the picture. Rather, a memory cache limited
to 4MB (~26 pictures) is used. When the memory cache is full, the last
accessed picture is remove from the cache and replaced with a new
picture on the top of the list. This avoids calling the server
unnecessarily, keeping performance maximal and avoiding fetching data to
the disk, which would not be that advantageous compared to obtaining
data from the web.

The Cache has been added under a MemoryCache package.

The UserCache has visibility to this new Memory Cache.

The UserMapper consults the new interface of the user cache in order to
see whether the profile picture is in the cache before calling the
gateways.