Changed icons in main menu to be translate-able
The icons in the main menu previously had the corresponding text as a part of the image, which meant that even translating the app in a different language (french, spanish, etc) the main menu icons would still display with the english text. Therefore now displaying corresponding text below the icon instead which can be altered by the app if the language of the user's device is a language other than english.
Also made the "my rides" button direct to the user ride information page. (MI-196)