MI-212 Initial implementation of the local SQLite database + unit tests
This initial implementation updates the project with the SQLite
component. It allows the application to store structured objects in the
a local database using object relational model.

The Cache package now contains a LocalDatabase package that defines the
location of the database, and that is responsible for creating the
tables when missing. The Cache also contains a new package named
LocalDataGateways. This new package contains gateways responsible for
communicating with the local database, the package is used by the logic
classes such as UserCache. The UserMapper will eventually communicate
with this UserCache in order to access the database indirectly, but this
is not implemented yet.

For the database to know how to define the table, the User model has
been updated with a primary key, and all the lists/relations with other
models have been ignored due to size limitation.

The SQLite and SQLiteAsync are classes added by the components handling
the SQLite database.

Unit tests for the UserLocalDataGateway has been added to test the
following functionalities: adding a user, deleting a user, finding a
user by id, updating a user. All four cases passed.