#28 Null check for list of reviews obtained from server added
For any reasons, if the server is either updating, or is down, the list
of feedback obtained may be null caused by an internal server error, or
any other reasons. The app side should robust enough to not crash
whenever that happens. A null check is added to the list before looping
through it