#106 Added code to communicate with the server in order to update a user's settings
- AbstractGateway:
Added the API endpoint link as a constant.

- UserGateway:
Added a method UpdateUserSettings(User) that will perform a post request
to the server to update the user's settings

UserMapper:
Modified the method UpdateUser to UpdateUserSettings, and update the
settings from the server by using the gateway.

PreferencesActivity:
Updated the name of UpdateUser to UpdateUserSettings